requirements.txt
beautifulsoup4
mysql-connector-python
requests
pyyaml
